% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg.R
\name{cg}
\alias{cg}
\title{cg}
\usage{
cg(colour1, colour2, n = 15)
}
\arguments{
\item{colour1}{The first hexcode colour.}

\item{colour2}{The second hexcode colour.}

\item{n}{The length of the vector returned by the function.}
}
\value{
A vector of hexcodes of length n, containing a colour gradient between colour =1 and colour2.
}
\description{
Creates colour gradient between two hexcodes.
}
\examples{
cg("white","black",20)

}
