% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalplaces.R
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{The number for the number of decimal places is to be measured.}
}
\value{
A single number, indicating the number of non-zero decimal places in `x`.
}
\description{
Tests the number of non-zero decimal places within a number.
}
\examples{
decimalplaces(23.43234525)
decimalplaces(334.3410000000000000)
decimalplaces(2.000)

}
